group-bot/requirements.txt
python-telegram-bot[job-queue]==20.3
openai>=1.0.0
python-dotenv